Enable scaled-dot for F4 inputs with F8 scales.
This change adds support for scaled-dot operations where the input operands are of type F4E2M1FN and the scales are of type F8E8M0FNU. It includes:
-   Adding a test case in JAX's scaled_dot_test.py for F4 types.
-   Adding a device test in Triton's fusion_emitter_device_test.cc to verify Triton's handling of F4 scaled-dot on Hopper GPUs.
-   Disabling cuBLAS autotuning for F4 types in gemm_fusion_autotuner.cc as cuBLAS does not support this.
-   Updating composite_rewriter.cc to recognize F4E2M1FN as a valid operand type for scaled-dot when paired with F8E8M0FNU scales.
-   Improved logging in composite_rewriter.cc for unsupported scaled-dot cases.

FP4 has an error in MMAv2 lowering path.
The line
auto dotOpA = cast<DotOperandEncodingAttr>(aTensorTy.getEncoding());
is wrong because the attr is not DotOperandEncodingAttr type.
as a result of that FP4 scaled dot lowering crashes on some tile sizes on B200 and always crash on H100.

PiperOrigin-RevId: 850440782